+++ front/src/components/CategoriaList/index.tsx
@@ -33,7 +33,7 @@ import { notify, TypeNotificationEnum } from '../../utils/notify';
 import {
   getCategoriaIcon,
   StatusEnum,
-} from '../MainDashboard/components/TransacaoModals/constants';
+} from '../../utils/constants';
 import { CreateCategoriaModal } from './components/CreateCategoriaModal';
 import { EditCategoriaModal } from './components/EditCategoriaModal';
