+++ front/src/hooks/useStorage/useStorage.ts
@@ -1,4 +1,5 @@
 import { useCallback, useState } from 'react';
+import { authVerify } from '../../services/auth';
 import storage from './storage';
 
 export default function useStorage(key: any) {
