+++ front/src/main.tsx
@@ -2,11 +2,10 @@ import { MantineProvider } from '@mantine/core';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClientProvider } from 'react-query';
+import { ReactQueryDevtools } from 'react-query/devtools';
 import App from './App';
-import AuthProvider from './context/AuthContext/AuthProvider';
 import { queryClient } from './services/queryClient';
 import { mantineTheme } from './utils/theme';
-import { ReactQueryDevtools } from 'react-query/devtools'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
@@ -16,11 +15,9 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
         withNormalizeCSS
         withGlobalStyles
       >
-        <AuthProvider>
-          <App />
-        </AuthProvider>
+        <App />
       </MantineProvider>
-    <ReactQueryDevtools />
+      <ReactQueryDevtools />
     </QueryClientProvider>
   </React.StrictMode>
 );
