+++ front/src/pages/CustomPage/index.tsx
@@ -21,19 +21,12 @@ import AuthContext from '../../context/AuthContext/AuthContext';
 import { findOneUser } from '../../services/user';
 
 export function CustomPage() {
-  const navigate = useNavigate();
   const { token } = useContext(AuthContext);
   const [openedUpdateName, handlersUpdateName] = useDisclosure(false);
 
-  const { data, isLoading, error } = useQuery(['user'], () => {
+  const { data } = useQuery(['user'], () => {
     return findOneUser(token.token);
   });
-
-  useEffect(() => {
-    if (!token) {
-      navigate('/login');
-    }
-  }, []);
 
   return (
     <Navigation>
