+++ front/src/pages/Home/index.tsx
@@ -6,15 +6,6 @@ import AuthContext from '../../context/AuthContext/AuthContext';
 import { MonthProvider } from '../../context/MonthContext/MonthContext';
 
 export function Home() {
-  const navigate = useNavigate();
-  const { token } = useContext(AuthContext);
-
-  useEffect(() => {
-    if (!token) {
-      navigate('/login');
-    }
-  }, []);
-
   return (
     <Navigation>
       <MonthProvider>
